# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the Compsole API documentation.\n\n### Authenticating\n\nThere are two methods of authenticating to the Compsole API. **Basic Auth** is used solely for the purpose of the Compsole UI. **Api Key Auth** is used for service accounts to authenticate prior to accessing the REST endpoints.\n\n#### Basic Auth\n\nUse the [`/auth/local/login`](#operations-Auth_API-post_auth_local_login) endpoint below to authenticate from the Compsole UI.\n\n#### Api Key Auth\n\nAPI Key Authentication is more complicated.\n\n1. You must retreive your `api_key` and `api_secret` from the Compsole UI after creating a service account.\n2. Use the [`/rest/token`](#operations-Auth_API-post_rest_login) endpoint to retrieve an API Token to use in requests\n3. Place the API Token into the `Authorization` header like so: `Authorization: Bearer \u003capi token here...\u003e`\n\n##### Refresh Tokens\n\nRefresh tokens can be used to renew a session without re-authenticating. The refresh token should be set in the `refresh-token` cookie already, so you can simply make a `POST` request to `/rest/token/refresh` and receive a new API token from this endpoint.\n",
        "title": "Compsole API",
        "contact": {
            "name": "BradHacker",
            "url": "http://github.com/BradHacker/compsole/issues"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
...
    },
    "basePath": "/api",
    "paths": {
...
        "/rest/user": {
            "get": {
                "security": [
                    {
                        "ServiceAuth": []
                    }
                ],
                "description": "List all Users",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Service API"
                ],
                "summary": "List all Users",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/rest.UserModel"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.APIError"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ServiceAuth": []
                    }
                ],
                "description": "Create a User",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Service API"
                ],
                "summary": "Create a User",
                "parameters": [
                    {
                        "description": "The user to create",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/rest.UserInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/rest.UserModel"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.APIError"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/api.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.APIError"
                        }
                    }
                }
            }
        },
        "/rest/user/{id}": {
            "get": {
                "security": [
                    {
                        "ServiceAuth": []
                    }
                ],
                "description": "Get a User",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Service API"
                ],
                "summary": "Get a User",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "example": "xxxxxxxx-xxxx-xxxx-xxxx-xxxxxxxxxxxx",
                        "description": "The id of the user",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/rest.UserModel"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.APIError"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/api.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.APIError"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ServiceAuth": []
                    }
                ],
                "description": "Update a User",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Service API"
                ],
                "summary": "Update a User",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "example": "xxxxxxxx-xxxx-xxxx-xxxx-xxxxxxxxxxxx",
                        "description": "The id of the user",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "The updated user",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/rest.UserInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/rest.UserModel"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.APIError"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/api.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.APIError"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ServiceAuth": []
                    }
                ],
                "description": "Delete a User",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Service API"
                ],
                "summary": "Delete a User",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "example": "xxxxxxxx-xxxx-xxxx-xxxx-xxxxxxxxxxxx",
                        "description": "The id of the user",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.APIError"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/api.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.APIError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "api.APIError": {
            "type": "object",
            "properties": {
                "error": {},
                "message": {
                    "type": "string"
                }
            }
        },
...
        "rest.TeamEdge": {
            "description": "Used for Team in edges",
            "type": "object",
            "properties": {
                "id": {
                    "description": "Fields",
                    "type": "string",
                    "example": "xxxxxxxx-xxxx-xxxx-xxxx-xxxxxxxxxxxx"
                },
                "name": {
                    "description": "[OPTIONAL] The display name for the team.",
                    "type": "string",
                    "example": "Team 1"
                },
                "team_number": {
                    "description": "[REQUIRED] The team number.",
                    "type": "integer",
                    "example": 1
                }
            }
        },
...
        "rest.UserInput": {
            "description": "Used as an input model for creating/updating Users",
            "type": "object",
            "required": [
                "first_name",
                "last_name",
                "role",
                "username"
            ],
            "properties": {
                "first_name": {
                    "type": "string",
                    "example": "John"
                },
                "last_name": {
                    "type": "string",
                    "example": "Doe"
                },
                "role": {
                    "type": "string",
                    "enum": [
                        "USER",
                        "ADMIN"
                    ],
                    "example": "USER"
                },
                "user_to_team": {
                    "type": "string",
                    "example": "xxxxxxxx-xxxx-xxxx-xxxx-xxxxxxxxxxxx"
                },
                "username": {
                    "type": "string",
                    "example": "compsole"
                }
            }
        },
        "rest.UserModel": {
            "description": "Used for User endpoints",
            "type": "object",
            "properties": {
                "first_name": {
                    "description": "[OPTIONAL] The display first name for the user.",
                    "type": "string",
                    "example": "Default"
                },
                "id": {
                    "description": "Fields",
                    "type": "string",
                    "example": "xxxxxxxx-xxxx-xxxx-xxxx-xxxxxxxxxxxx"
                },
                "last_name": {
                    "description": "[OPTIONAL] The display last name for the user.",
                    "type": "string",
                    "example": "User"
                },
                "role": {
                    "description": "[REQUIRED] The role of the user. Admins have full access.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/user.Role"
                        }
                    ],
                    "example": "USER"
                },
                "user_to_team": {
                    "description": "Edges",
                    "allOf": [
                        {
                            "$ref": "#/definitions/rest.TeamEdge"
                        }
                    ]
                },
                "username": {
                    "description": "[REQUIRED] The username for the user.",
                    "type": "string",
                    "example": "compsole"
                }
            }
        },
...
        "user.Role": {
            "type": "string",
            "enum": [
                "USER",
                "ADMIN"
            ],
            "x-enum-varnames": [
                "RoleUSER",
                "RoleADMIN"
            ]
        }
    },
    "securityDefinitions": {
        "ServiceAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "UserAuth": {
            "type": "basic"
        }
    },
...
}